Grammar upgrade
'Multiplication' is now 'MultDiv' and does both multiplication and division
'Addition' is now 'AddSub' and does both addition and subtraction

Added 'stmt', 'stmts' & 'cond'
'stmt' contains various statements which are all labeled: Assignment, If, IfElse & While
'stmts' contains the ability to chain several 'stmt's together. This makes it easier to form the body of a while loop for example.
'cond' contains various conditionals:
equals/not equals : `==`/`!=`
less than/greater than: `<`/`>`
and/or: `&&`/`||`
not: `!`

Java part:
"implemented" the methods required for the parser to accept the new grammar, but the actual implementation of the methods have not yet been done
Also changed the `visitStart` method to work with statements instead of the previous assignments
:ok_hand: